pyproject.toml
[tool.poetry]
name = "gym-chess"
version = "0.1.1"
description = "OpenAI Gym environments for Chess"
authors = ["Lucas Wolf <iamlucaswolf@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/iamlucaswolf/gym-chess"

[tool.poetry.dependencies]
python = "^3.6"
python-chess = "^0.31.1"
gym = "^0.17.2"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
vermin = "^0.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
